+++ CHANGES.rst
@@ -4,5 +4,9 @@ Changelog
 0.3 (unreleased)
 ----------------
 
+- Restructure blueprints to be in blueprints directory and integrate orphaned
+  blueprints from collective.blueprint.jsonmigrator.
+  [thet]
+
 - PEP 8.
   [thet]